change distance functions to be floats

+++ nd4j-api/src/main/java/org/nd4j/linalg/distancefunction/CosineSimilarity.java
@@ -14,8 +14,8 @@ public class CosineSimilarity extends BaseDistanceFunction {
 	}
 
 	@Override
-	public Double apply(INDArray input) {
-		return Transforms.cosineSim(input, base);
+	public Float apply(INDArray input) {
+		return (float) Transforms.cosineSim(input, base);
 	}
